LOCL is long (64 vs 32 bit stuff)
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@586 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/networkTest/networkTest.cpp
@@ -28,7 +28,7 @@ char             host[80];
 
 
 
-#define LOCL (int) "localhost"
+#define LOCL (long) "localhost"
 /* ****************************************************************************
 *
 * parse arguments
